refactor: EventBridgeWriter
Mapping from Kafkas `SinkRecord` to required Event Bridge
`PutEventsRequestEntry` was extracted to interface `EventBridgeMapper` with a
Jackson implementation by `DefaultEventBridgeMapper`. Also batching was
decoupled by introducing the functional interface `EventBridgeBatchingStrategy`
with the current simple implementation of single item per batch.

The state of a mapped `SinkRecord` changed from a mutable (data) class to the
immutable `EventBridgeResult` which holds either a `SinkRecord` and a
succesfull mapping or a `SinkRecord` with an dedicated error.

Fixes #32, fixes #46
Closes #35, closes #47

Co-authored-by: maschnetwork <maxschellhorn@pm.me>