feat: add AWS S3 sidelining (aka claim check)
Add feature to extract and send (sideline) EventBridge event value, which is
the content of the Kafka record value, to S3 by a given JSON Path. The event
with the transformed value is send to EventBridge with the reference to the S3
object and the used JSON Path.

The feature is activated if the S3 bucket (e.g. 'my-bucket') is configured for
the connector by:
```json5
{
  "config": {
    // other required configuration
    "aws.eventbridge.offloading.s3.default.bucket": "my-bucket"
  }
}
```

The user must have write access to the configured bucket.

The default JSON Path is `$.detail.value` referencing a EventBridge entry. A
valid JSON Path must start with `$.detail.value` and must reference a single
JSON object, array or value.

The JSON Path can be configured for the connector by:
```json5
{
  "config": {
    // other required configuration
    "aws.eventbridge.offloading.default.fieldref": "$.detail.value"
  }
}
```

If the JSON Path does not match, nothing is send to S3. Otherwise, if the JSON
Path matches a non `null` JSON object, array or value, this value is send to S3
and the JSON attribute is removed. If the value is `null` and matches the JSON
Path, the JSON attribute with it's `null` value is keept and nothing send to
S3.

If the feature is activated and a value was send to S3, the EventBridge event
contains the attribute `$.detail.datarefJsonPath` with the configured JSON Path
and the S3 object reference at `$.detail.dataref`. Otherwise the EventBridge
event is unchanged.

A FIFO cache with a maximum size of 10.000 is used to look the uploaded content
to S3 by it's SHA512 and re-use the S3 object key whenever possible.

Closes #110.